+++ hutool-captcha/src/main/java/cn/hutool/captcha/CircleCaptcha.java
@@ -118,6 +118,11 @@ public class CircleCaptcha extends AbstractCaptcha {
 	 * @param g {@link Graphics2D}
 	 */
 	private void drawInterfere(Graphics2D g) {
+		// issue#IDJQ15 自定义线条特征（粗细等）
+		if(null != this.stroke){
+			g.setStroke(this.stroke);
+		}
+
 		final ThreadLocalRandom random = RandomUtil.getRandom();
 
 		for (int i = 0; i < this.interfereCount; i++) {